requirements.txt
colorama==0.4.6
discord.py==2.3.2
discord_webhook==1.3.0
python-dotenv==1.0.0
asyncio==3.4.3